restructured the files, everything should be in its proper place now
